only require wasm dependencies when targetting wasm
